+++ internal/provider/waypoint/data_source_waypoint_application.go
@@ -68,17 +68,37 @@ func (d *DataSourceApplication) Schema(ctx context.Context, req datasource.Schem
 				Computed:    true,
 				Description: "Instructions for using the Application (markdown format supported).",
 			},
-			"application_template_id": schema.StringAttribute{
+			"template_id": schema.StringAttribute{
 				Computed:    true,
-				Description: "ID of the Application Template this Application is based on.",
+				Description: "ID of the Template this Application is based on.",
 			},
-			"application_template_name": schema.StringAttribute{
+			"template_name": schema.StringAttribute{
 				Computed:    true,
-				Description: "Name of the Application Template this Application is based on.",
+				Description: "Name of the Template this Application is based on.",
 			},
 			"namespace_id": schema.StringAttribute{
 				Computed:    true,
 				Description: "Internal Namespace ID.",
+			},
+			"input_variables": schema.SetNestedAttribute{
+				Optional:    true,
+				Description: "Input variables for the Application.",
+				NestedObject: schema.NestedAttributeObject{
+					Attributes: map[string]schema.Attribute{
+						"name": &schema.StringAttribute{
+							Computed:    true,
+							Description: "Variable name",
+						},
+						"value": &schema.StringAttribute{
+							Computed:    true,
+							Description: "Variable value",
+						},
+						"variable_type": &schema.StringAttribute{
+							Computed:    true,
+							Description: "Variable type",
+						},
+					},
+				},
 			},
 		},
 	}
@@ -100,8 +120,28 @@ func (d *DataSourceApplication) Configure(ctx context.Context, req datasource.Co
 	d.client = client
 }
 
+// ApplicationDataSourceModel describes the data source data model
+type ApplicationDataSourceModel struct {
+	ID             types.String `tfsdk:"id"`
+	Name           types.String `tfsdk:"name"`
+	ProjectID      types.String `tfsdk:"project_id"`
+	OrgID          types.String `tfsdk:"organization_id"`
+	ReadmeMarkdown types.String `tfsdk:"readme_markdown"`
+	TemplateID     types.String `tfsdk:"template_id"`
+	TemplateName   types.String `tfsdk:"template_name"`
+	NamespaceID    types.String `tfsdk:"namespace_id"`
+
+	// deferred for now
+	// Tags       types.List `tfsdk:"tags"`
+
+	// deferred and probably a list or objects, but may possible be a separate
+	// ActionCfgs types.List `tfsdk:"action_cfgs"`
+
+	InputVars types.Set `tfsdk:"input_variables"`
+}
+
 func (d *DataSourceApplication) Read(ctx context.Context, req datasource.ReadRequest, resp *datasource.ReadResponse) {
-	var data ApplicationResourceModel
+	var data ApplicationDataSourceModel
 	resp.Diagnostics.Append(req.Config.Get(ctx, &data)...)
 
 	client := d.client
@@ -144,6 +184,31 @@ func (d *DataSourceApplication) Read(ctx context.Context, req datasource.ReadReq
 	data.ReadmeMarkdown = types.StringValue(application.ReadmeMarkdown.String())
 	if application.ReadmeMarkdown.String() == "" {
 		data.ReadmeMarkdown = types.StringNull()
+	}
+
+	// A second API call is made to get the input vars set on the application
+	inputVars, err := clients.GetInputVariables(ctx, client, data.Name.ValueString(), loc)
+	if err != nil {
+		resp.Diagnostics.AddError(err.Error(), "Failed to fetch application's input variables.")
+		return
+	}
+
+	inputVariables := make([]*InputVar, 0)
+	for _, iv := range inputVars {
+		inputVariables = append(inputVariables, &InputVar{
+			Name:  types.StringValue(iv.Name),
+			Value: types.StringValue(iv.Value),
+		})
+	}
+	if len(inputVariables) > 0 {
+		aivs, diags := types.SetValueFrom(ctx, types.ObjectType{AttrTypes: InputVar{}.attrTypes()}, inputVariables)
+		resp.Diagnostics.Append(diags...)
+		if diags.HasError() {
+			return
+		}
+		data.InputVars = aivs
+	} else {
+		data.InputVars = types.SetNull(types.ObjectType{AttrTypes: InputVar{}.attrTypes()})
 	}
 
 	resp.Diagnostics.Append(resp.State.Set(ctx, &data)...)