+++ internal/provider/waypoint/resource_waypoint_template.go
@@ -11,6 +11,7 @@ import (
 	sharedmodels "github.com/hashicorp/hcp-sdk-go/clients/cloud-shared/v1/models"
 	"github.com/hashicorp/hcp-sdk-go/clients/cloud-waypoint-service/preview/2023-08-18/client/waypoint_service"
 	waypoint_models "github.com/hashicorp/hcp-sdk-go/clients/cloud-waypoint-service/preview/2023-08-18/models"
+	"github.com/hashicorp/terraform-plugin-framework-validators/listvalidator"
 	"github.com/hashicorp/terraform-plugin-framework/diag"
 	"github.com/hashicorp/terraform-plugin-framework/path"
 	"github.com/hashicorp/terraform-plugin-framework/resource"
@@ -18,26 +19,27 @@ import (
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema/listplanmodifier"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema/planmodifier"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema/stringplanmodifier"
+	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
 	"github.com/hashicorp/terraform-plugin-framework/types"
 	"github.com/hashicorp/terraform-plugin-log/tflog"
 	"github.com/hashicorp/terraform-provider-hcp/internal/clients"
 )
 
 // Ensure provider defined types fully satisfy framework interfaces.
-var _ resource.Resource = &ApplicationTemplateResource{}
-var _ resource.ResourceWithImportState = &ApplicationTemplateResource{}
+var _ resource.Resource = &TemplateResource{}
+var _ resource.ResourceWithImportState = &TemplateResource{}
 
-func NewApplicationTemplateResource() resource.Resource {
-	return &ApplicationTemplateResource{}
+func NewTemplateResource() resource.Resource {
+	return &TemplateResource{}
 }
 
-// ApplicationTemplateResource defines the resource implementation.
-type ApplicationTemplateResource struct {
+// TemplateResource defines the resource implementation.
+type TemplateResource struct {
 	client *clients.Client
 }
 
-// ApplicationTemplateResourceModel describes the resource data model.
-type ApplicationTemplateResourceModel struct {
+// TemplateResourceModel describes the resource data model.
+type TemplateResourceModel struct {
 	ID                     types.String `tfsdk:"id"`
 	Name                   types.String `tfsdk:"name"`
 	ProjectID              types.String `tfsdk:"project_id"`
@@ -70,36 +72,36 @@ type tfcVariableOption struct {
 	UserEditable types.Bool   `tfsdk:"user_editable"`
 }
 
-func (r *ApplicationTemplateResource) Metadata(ctx context.Context, req resource.MetadataRequest, resp *resource.MetadataResponse) {
-	resp.TypeName = req.ProviderTypeName + "_waypoint_application_template"
+func (r *TemplateResource) Metadata(ctx context.Context, req resource.MetadataRequest, resp *resource.MetadataResponse) {
+	resp.TypeName = req.ProviderTypeName + "_waypoint_template"
 }
 
-func (r *ApplicationTemplateResource) Schema(ctx context.Context, req resource.SchemaRequest, resp *resource.SchemaResponse) {
+func (r *TemplateResource) Schema(ctx context.Context, req resource.SchemaRequest, resp *resource.SchemaResponse) {
 	resp.Schema = schema.Schema{
 		// This description is used by the documentation generator and the language server.
-		MarkdownDescription: "Waypoint Application Template resource",
+		MarkdownDescription: "Waypoint Template resource",
 
 		Attributes: map[string]schema.Attribute{
 			"id": schema.StringAttribute{
 				Computed:    true,
-				Description: "The ID of the Application Template.",
+				Description: "The ID of the Template.",
 				PlanModifiers: []planmodifier.String{
 					stringplanmodifier.UseStateForUnknown(),
 				},
 			},
 			"name": schema.StringAttribute{
-				Description: "The name of the Application Template.",
+				Description: "The name of the Template.",
 				Required:    true,
 			},
 			"organization_id": schema.StringAttribute{
-				Description: "The ID of the HCP organization where the Waypoint Application Template is located.",
+				Description: "The ID of the HCP organization where the Waypoint Template is located.",
 				Computed:    true,
 				PlanModifiers: []planmodifier.String{
 					stringplanmodifier.UseStateForUnknown(),
 				},
 			},
 			"project_id": schema.StringAttribute{
-				Description: "The ID of the HCP project where the Waypoint Application Template is located.",
+				Description: "The ID of the HCP project where the Waypoint Template is located.",
 				Optional:    true,
 				Computed:    true,
 				PlanModifiers: []planmodifier.String{
@@ -121,7 +123,7 @@ func (r *ApplicationTemplateResource) Schema(ctx context.Context, req resource.S
 			"labels": schema.ListAttribute{
 				// Computed:    true,
 				Optional:    true,
-				Description: "List of labels attached to this Application Template.",
+				Description: "List of labels attached to this Template.",
 				ElementType: types.StringType,
 				PlanModifiers: []planmodifier.List{
 					listplanmodifier.UseStateForUnknown(),
@@ -169,8 +171,12 @@ func (r *ApplicationTemplateResource) Schema(ctx context.Context, req resource.S
 							Description: "Variable type",
 						},
 						"options": &schema.ListAttribute{
+							Optional:    true,
+							Computed:    true,
 							ElementType: types.StringType,
-							Required:    true,
+							Validators: []validator.List{
+								listvalidator.UniqueValues(),
+							},
 							Description: "List of options",
 						},
 						"user_editable": &schema.BoolAttribute{
@@ -186,7 +192,7 @@ func (r *ApplicationTemplateResource) Schema(ctx context.Context, req resource.S
 	}
 }
 
-func (r *ApplicationTemplateResource) Configure(ctx context.Context, req resource.ConfigureRequest, resp *resource.ConfigureResponse) {
+func (r *TemplateResource) Configure(ctx context.Context, req resource.ConfigureRequest, resp *resource.ConfigureResponse) {
 	// Prevent panic if the provider has not been configured.
 	if req.ProviderData == nil {
 		return
@@ -205,8 +211,8 @@ func (r *ApplicationTemplateResource) Configure(ctx context.Context, req resourc
 	r.client = client
 }
 
-func (r *ApplicationTemplateResource) Create(ctx context.Context, req resource.CreateRequest, resp *resource.CreateResponse) {
-	var plan *ApplicationTemplateResourceModel
+func (r *TemplateResource) Create(ctx context.Context, req resource.CreateRequest, resp *resource.CreateResponse) {
+	var plan *TemplateResourceModel
 
 	// Read Terraform plan data into the model
 	resp.Diagnostics.Append(req.Plan.Get(ctx, &plan)...)
@@ -291,7 +297,7 @@ func (r *ApplicationTemplateResource) Create(ctx context.Context, req resource.C
 	}
 	createTplResp, err := r.client.Waypoint.WaypointServiceCreateApplicationTemplate(params, nil)
 	if err != nil {
-		resp.Diagnostics.AddError("Error creating application template", err.Error())
+		resp.Diagnostics.AddError("Error creating template", err.Error())
 		return
 	}
 
@@ -300,7 +306,7 @@ func (r *ApplicationTemplateResource) Create(ctx context.Context, req resource.C
 		appTemplate = createTplResp.Payload.ApplicationTemplate
 	}
 	if appTemplate == nil {
-		resp.Diagnostics.AddError("unknown error creating application template", "empty application template returned")
+		resp.Diagnostics.AddError("unknown error creating template", "empty template returned")
 		return
 	}
 
@@ -358,7 +364,7 @@ func (r *ApplicationTemplateResource) Create(ctx context.Context, req resource.C
 
 	// Write logs using the tflog package
 	// Documentation: https://terraform.io/plugin/log
-	tflog.Trace(ctx, "created application template resource")
+	tflog.Trace(ctx, "created template resource")
 
 	// Save data into Terraform state
 	resp.Diagnostics.Append(resp.State.Set(ctx, &plan)...)
@@ -392,8 +398,8 @@ func readVarOpts(
 	return varOpts, nil
 }
 
-func (r *ApplicationTemplateResource) Read(ctx context.Context, req resource.ReadRequest, resp *resource.ReadResponse) {
-	var data *ApplicationTemplateResourceModel
+func (r *TemplateResource) Read(ctx context.Context, req resource.ReadRequest, resp *resource.ReadResponse) {
+	var data *TemplateResourceModel
 
 	// Read Terraform prior state data into the model
 	resp.Diagnostics.Append(req.State.Get(ctx, &data)...)
@@ -480,8 +486,8 @@ func (r *ApplicationTemplateResource) Read(ctx context.Context, req resource.Rea
 	resp.Diagnostics.Append(resp.State.Set(ctx, &data)...)
 }
 
-func (r *ApplicationTemplateResource) Update(ctx context.Context, req resource.UpdateRequest, resp *resource.UpdateResponse) {
-	var plan *ApplicationTemplateResourceModel
+func (r *TemplateResource) Update(ctx context.Context, req resource.UpdateRequest, resp *resource.UpdateResponse) {
+	var plan *TemplateResourceModel
 
 	// Read Terraform plan data into the model
 	resp.Diagnostics.Append(req.Plan.Get(ctx, &plan)...)
@@ -577,7 +583,7 @@ func (r *ApplicationTemplateResource) Update(ctx context.Context, req resource.U
 		appTemplate = app.Payload.ApplicationTemplate
 	}
 	if appTemplate == nil {
-		resp.Diagnostics.AddError("unknown error updating application template", "empty application template returned")
+		resp.Diagnostics.AddError("unknown error updating template", "empty template returned")
 		return
 	}
 
@@ -627,14 +633,14 @@ func (r *ApplicationTemplateResource) Update(ctx context.Context, req resource.U
 
 	// Write logs using the tflog package
 	// Documentation: https://terraform.io/plugin/log
-	tflog.Trace(ctx, "updated application template resource")
+	tflog.Trace(ctx, "updated template resource")
 
 	// Save updated data into Terraform state
 	resp.Diagnostics.Append(resp.State.Set(ctx, &plan)...)
 }
 
-func (r *ApplicationTemplateResource) Delete(ctx context.Context, req resource.DeleteRequest, resp *resource.DeleteResponse) {
-	var data *ApplicationTemplateResourceModel
+func (r *TemplateResource) Delete(ctx context.Context, req resource.DeleteRequest, resp *resource.DeleteResponse) {
+	var data *TemplateResourceModel
 
 	// Read Terraform prior state data into the model
 	resp.Diagnostics.Append(req.State.Get(ctx, &data)...)
@@ -671,17 +677,17 @@ func (r *ApplicationTemplateResource) Delete(ctx context.Context, req resource.D
 	_, err = r.client.Waypoint.WaypointServiceDeleteApplicationTemplate(params, nil)
 	if err != nil {
 		if clients.IsResponseCodeNotFound(err) {
-			tflog.Info(ctx, "Application Template not found for organization during delete call, ignoring")
+			tflog.Info(ctx, "Template not found for organization during delete call, ignoring")
 			return
 		}
 		resp.Diagnostics.AddError(
-			"Error Deleting Application Template",
+			"Error Deleting Template",
 			err.Error(),
 		)
 		return
 	}
 }
 
-func (r *ApplicationTemplateResource) ImportState(ctx context.Context, req resource.ImportStateRequest, resp *resource.ImportStateResponse) {
+func (r *TemplateResource) ImportState(ctx context.Context, req resource.ImportStateRequest, resp *resource.ImportStateResponse) {
 	resource.ImportStatePassthroughID(ctx, path.Root("id"), req, resp)
 }